proj/requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.7
billiard==3.5.0.5
celery
Django==3.0.5
django-appconf==1.0.4
django-celery==3.3.1
django-celery-email==3.0.0
django-celery-results==1.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
redis
psycopg2
